plc-service/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
python-snap7>=1.3
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
apscheduler>=3.10.0
